refactor: Restructure internal tools and improve search functionality
- Moved MCPSearch and MCPInstall to a new internal-tools directory
- Added support for exact search matching
- Introduced more flexible tool injection and initialization
- Enhanced tool metadata with categories and config details
- Simplified tool loading and search logic in index.ts
